+++ presqt/api_v1/utilities/__init__.py
@@ -1,8 +1,4 @@
 from presqt.api_v1.utilities.fixity.hash_generator import hash_generator
-from presqt.api_v1.utilities.io.read_file import read_file
-from presqt.api_v1.utilities.io.write_file import write_file
-from presqt.api_v1.utilities.io.zip_file import zip_directory
-from presqt.api_v1.utilities.utils.compare_lists import  compare_lists
 from presqt.api_v1.utilities.utils.get_target_data import get_target_data
 from presqt.api_v1.utilities.utils.function_router import FunctionRouter
 from presqt.api_v1.utilities.utils.target_actions import (
@@ -17,17 +13,13 @@ from presqt.api_v1.utilities.validation.token_validation import (get_source_toke
 
 __all__ = [
     action_checker,
-    compare_lists,
     link_builder,
-    write_file,
-    read_file,
-    zip_directory,
+    hash_generator,
     FunctionRouter,
     target_validation,
     get_source_token,
     get_destination_token,
     get_target_data,
-    hash_generator,
     file_duplicate_action_validation,
     get_process_info_data,
     process_token_validation
