+++ presqt/api_v1/utilities/validation/file_validation.py
@@ -3,7 +3,7 @@ import zipfile
 from django.utils.datastructures import MultiValueDictKeyError
 from rest_framework import status
 
-from presqt.exceptions import PresQTValidationError
+from presqt.utilities import PresQTValidationError
 
 
 def file_validation(request):
