+++ presqt/targets/osf/classes/base.py
@@ -1,8 +1,8 @@
 from rest_framework import status
 
-from presqt.exceptions import PresQTInvalidTokenError, PresQTResponseException
-from presqt.osf.exceptions import OSFNotFoundError, OSFForbiddenError
-from presqt.session import PresQTSession
+from presqt.utilities import PresQTInvalidTokenError, PresQTResponseException
+from presqt.targets.osf.exceptions import OSFNotFoundError, OSFForbiddenError
+from presqt.targets.utilities.session import PresQTSession
 
 
 class OSFBase(object):
