+++ presqt/targets/osf/classes/storage_folder.py
@@ -3,10 +3,11 @@ import os
 from requests.exceptions import ConnectionError
 from rest_framework import status
 
-from presqt.api_v1.utilities import read_file, hash_generator
-from presqt.exceptions import PresQTResponseException
-from presqt.osf.classes.base import OSFBase
-from presqt.osf.classes.file import File
+from presqt.api_v1.utilities import hash_generator
+from presqt.utilities import read_file
+from presqt.utilities import PresQTResponseException
+from presqt.targets.osf.classes.base import OSFBase
+from presqt.targets.osf.classes.file import File
 
 class ContainerMixin:
     def get_resources_objects(self, container):
