+++ presqt/utilities/io/remove_path_contents.py
